Added longer

+++ src/Main.java
@@ -2,7 +2,7 @@ package src;
 
 public class Main {
     public static void main(String[] args) {
-        int[] price = {20, 30, 10};
+        long[] price = {20, 30, 10};
 
         SalesManager sales = new SalesManager(price);
         System.out.println("Sale max = " + sales.max());
